requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
Django==4.2.2
django-autoslug==1.9.9
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
djangorestframework==3.14.0
docutils==0.20.1
Faker==19.1.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
gunicorn==21.2.0
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
kombu==5.3.1
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
packaging==23.1
pguard==0.4.0
Pillow==9.5.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-memcached==1.59
pytz==2023.3
redis==4.6.0
requests==2.31.0
setuptools-scm==7.1.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==7.1.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
